🎨 🚨 Split theme tests, clean config & add tests
refs #7491

- split themes_spec up into several files
- clean up the code for configuration
- ensure its tested
