+++ core/server/themes/active.js
@@ -18,24 +18,12 @@
  */
 var _ = require('lodash'),
     join = require('path').join,
-    defaultConfig = require('./defaults.json'),
+    themeConfig = require('./config'),
     config = require('../config'),
     // @TODO: remove this require
     hbs = require('express-hbs'),
     // Current instance of ActiveTheme
     currentActiveTheme;
-
-// @TODO: will clean this code up later, honest! (and add tests)
-function tempConfigHandler(packageJson) {
-    var config = _.cloneDeep(defaultConfig),
-        allowedKeys = ['posts_per_page'];
-
-    if (packageJson && packageJson.hasOwnProperty('config')) {
-        config = _.assign(config, _.pick(packageJson.config, allowedKeys));
-    }
-
-    return config;
-}
 
 class ActiveTheme {
     /**
@@ -61,8 +49,8 @@ class ActiveTheme {
             return templates;
         }, []);
 
-        // Do something with config here
-        this._config = tempConfigHandler(this._packageInfo);
+        // Create a theme config object
+        this._config = themeConfig.create(this._packageInfo);
     }
 
     get name() {
