+++ core/server/themes/config/defaults.json
